Disable coverage "can't parse" warnings
When running coverage using Python < 3.10, we don't want to see the
warnings about being unable to parse code that uses the `match`
construct introduce in Python 3.10.

In addition, remove `--cov=tests` because coverage is meant to measure
only how much of the _main_ code is covered during test execution, not
how much _test_ code is covered.

Lastely, change `term` to `term-missing` to identify _exactly_ which
lines of code are not covered by tests.